[wasm] Extract a base class from WasmAppBuilder, for future use with WasiAppBuilder task
