+++ src/tasks/WasmAppBuilder/WasmAppBuilder.cs
@@ -13,38 +13,16 @@ using System.Text.Json.Serialization;
 using Microsoft.Build.Framework;
 using Microsoft.Build.Utilities;
 
-public class WasmAppBuilder : Task
-{
-    [NotNull]
-    [Required]
-    public string? AppDir { get; set; }
+namespace Microsoft.WebAssembly.Build.Tasks;
 
+public class WasmAppBuilder : WasmAppBuilderBaseTask
+{
     [NotNull]
     [Required]
     public string? MainJS { get; set; }
 
-    [Required]
-    public string[] Assemblies { get; set; } = Array.Empty<string>();
-
-    [NotNull]
-    [Required]
-    public ITaskItem[]? NativeAssets { get; set; }
-
-    private readonly List<string> _fileWrites = new();
-
-    [Output]
-    public string[]? FileWrites => _fileWrites.ToArray();
-
-    // full list of ICU data files we produce can be found here:
-    // https://github.com/dotnet/icu/tree/maint/maint-67/icu-filters
-    public string? IcuDataFileName { get; set; }
-
-    public int DebugLevel { get; set; }
-    public ITaskItem[]? SatelliteAssemblies { get; set; }
     public ITaskItem[]? FilesToIncludeInFileSystem { get; set; }
     public ITaskItem[]? RemoteSources { get; set; }
-    public bool InvariantGlobalization { get; set; }
-    public ITaskItem[]? ExtraFilesToDeploy { get; set; }
     public string? MainHTMLPath { get; set; }
     public bool IncludeThreadsWorker {get; set; }
     public int PThreadPoolSize {get; set; }
@@ -63,16 +41,6 @@ public class WasmAppBuilder : Task
     //       <WasmExtraConfig Include="string_with_json" Value="&quot;{ &quot;abc&quot;: 4 }&quot;" />
     // </summary>
     public ITaskItem[]? ExtraConfig { get; set; }
-
-    public string? DefaultHostConfig { get; set; }
-
-    [Required, NotNull]
-    public string? MainAssemblyName { get; set; }
-
-    [Required]
-    public ITaskItem[] HostConfigs { get; set; } = Array.Empty<ITaskItem>();
-
-    public ITaskItem[] RuntimeArgsForHost { get; set; } = Array.Empty<ITaskItem>();
 
     private sealed class WasmAppConfig
     {
@@ -150,20 +118,7 @@ public class WasmAppBuilder : Task
         public bool LoadRemote { get; set; }
     }
 
-    public override bool Execute ()
-    {
-        try
-        {
-            return ExecuteInternal();
-        }
-        catch (LogAsErrorException laee)
-        {
-            Log.LogError(laee.Message);
-            return false;
-        }
-    }
-
-    private bool ExecuteInternal ()
+    protected override bool ExecuteInternal()
     {
         if (!File.Exists(MainJS))
             throw new LogAsErrorException($"File MainJS='{MainJS}' doesn't exist.");
@@ -425,77 +380,6 @@ public class WasmAppBuilder : Task
         return !Log.HasLoggedErrors;
     }
 
-    private void UpdateRuntimeConfigJson()
-    {
-        string[] matchingAssemblies = Assemblies.Where(asm => Path.GetFileName(asm) == MainAssemblyName).ToArray();
-        if (matchingAssemblies.Length == 0)
-            throw new LogAsErrorException($"Could not find main assembly named {MainAssemblyName} in the list of assemblies");
-
-        if (matchingAssemblies.Length > 1)
-            throw new LogAsErrorException($"Found more than one assembly matching the main assembly name {MainAssemblyName}: {string.Join(",", matchingAssemblies)}");
-
-        string runtimeConfigPath = Path.ChangeExtension(matchingAssemblies[0], ".runtimeconfig.json");
-        if (!File.Exists(runtimeConfigPath))
-        {
-            Log.LogMessage(MessageImportance.Low, $"Could not find {runtimeConfigPath}. Ignoring.");
-            return;
-        }
-
-        var rootNode = JsonNode.Parse(File.ReadAllText(runtimeConfigPath),
-                                            new JsonNodeOptions { PropertyNameCaseInsensitive = true });
-        if (rootNode == null)
-            throw new LogAsErrorException($"Failed to parse {runtimeConfigPath}");
-
-        JsonObject? rootObject = rootNode.AsObject();
-        if (!rootObject.TryGetPropertyValue("runtimeOptions", out JsonNode? runtimeOptionsNode)
-                || !(runtimeOptionsNode is JsonObject runtimeOptionsObject))
-        {
-            throw new LogAsErrorException($"Could not find node named 'runtimeOptions' in {runtimeConfigPath}");
-        }
-
-        JsonObject wasmHostProperties = runtimeOptionsObject.GetOrCreate<JsonObject>("wasmHostProperties", () => new JsonObject());
-        JsonArray runtimeArgsArray = wasmHostProperties.GetOrCreate<JsonArray>("runtimeArgs", () => new JsonArray());
-        JsonArray perHostConfigs = wasmHostProperties.GetOrCreate<JsonArray>("perHostConfig", () => new JsonArray());
-
-        if (string.IsNullOrEmpty(DefaultHostConfig) && HostConfigs.Length > 0)
-            DefaultHostConfig = HostConfigs[0].ItemSpec;
-
-        if (!string.IsNullOrEmpty(DefaultHostConfig))
-            wasmHostProperties["defaultConfig"] = DefaultHostConfig;
-
-        wasmHostProperties["mainAssembly"] = MainAssemblyName;
-
-        foreach (JsonValue? rarg in RuntimeArgsForHost.Select(ri => JsonValue.Create(ri.ItemSpec)))
-        {
-            if (rarg is not null)
-                runtimeArgsArray.Add(rarg);
-        }
-
-        foreach (ITaskItem hostConfigItem in HostConfigs)
-        {
-            var hostConfigObject = new JsonObject();
-
-            string name = hostConfigItem.ItemSpec;
-            string host = hostConfigItem.GetMetadata("host");
-            if (string.IsNullOrEmpty(host))
-                throw new LogAsErrorException($"BUG: Could not find required metadata 'host' for host config named '{name}'");
-
-            hostConfigObject.Add("name", name);
-            foreach (KeyValuePair<string, string> kvp in hostConfigItem.CloneCustomMetadata().Cast<KeyValuePair<string, string>>())
-                hostConfigObject.Add(kvp.Key, kvp.Value);
-
-            perHostConfigs.Add(hostConfigObject);
-        }
-
-        string dstPath = Path.Combine(AppDir!, Path.GetFileName(runtimeConfigPath));
-        using FileStream? fs = File.OpenWrite(dstPath);
-        using var writer = new Utf8JsonWriter(fs, new JsonWriterOptions { Indented = true });
-        rootObject.WriteTo(writer);
-        _fileWrites.Add(dstPath);
-
-        Log.LogMessage(MessageImportance.Low, $"Generated {dstPath} from {runtimeConfigPath}");
-    }
-
     private bool TryParseExtraConfigValue(ITaskItem extraItem, out object? valueObject)
     {
         valueObject = null;
@@ -538,28 +422,6 @@ public class WasmAppBuilder : Task
         catch (Exception ex) when (ex is FormatException or InvalidCastException or OverflowException)
         {
             return false;
-        }
-    }
-
-    private bool FileCopyChecked(string src, string dst, string label)
-    {
-        if (!File.Exists(src))
-        {
-            Log.LogError($"{label} file '{src}' not found");
-            return false;
-        }
-
-        Log.LogMessage(MessageImportance.Low, $"Copying file from '{src}' to '{dst}'");
-        try
-        {
-            File.Copy(src, dst, true);
-            _fileWrites.Add(dst);
-
-            return true;
-        }
-        catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
-        {
-            throw new LogAsErrorException($"{label} Failed to copy {src} to {dst} because {ex.Message}");
         }
     }
 }
